moved LHE filter to GeneratorInterface/GenFilters
